# config/clubSwagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "API",
		"description": "Documentation",
		"version": "1.0"
	},
	"produces": ["application/json"],
	"basePath": "/",
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	},
	"paths": {
...
		"/api/v1/deleteEvent": {
			"delete": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "deleteEvent",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Application Club"],
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "formData",
						"description": "Enter the id of event",
						"required": true,
						"type": "string",
						"value": ""
					}
				],
				"responses": {}
			}
		},
		"/api/v1/myCreatedEvent": {
			"get": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "myCreatedEvent",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Application Club"],
				"description": "",
				"parameters": [
				],
				"responses": {}
			}
		},
		"/api/v1/clubAllEvent": {
			"post": {
				"x-swagger-router-controller": "middleware-name1",
				"operationId": "clubAllEvent",
				"security": [
					{
						"Bearer": []
					}
				],
				"tags": ["Application Club"],
				"description": "",
				"parameters": [
					{
						"name": "clubId",
						"in": "formData",
						"description": "Enter the clubId",
						"required": true,
						"type": "string",
						"value": ""
					}
				],
				"responses": {}
			}
		},
...
	},
...
}